requirements.txt
Django
djangorestframework
Pillow
django-filter>=23.0